Allow Terragrunt compatibility
The main branch of this module and Terragrunt do not play well together
because the `*.tfvars` files are interpreted with the module code delivered
in the respective directories, creating a conflict in the argument values
delivered to the module.

By commenting out the `*.tfvars` files *in their entirety*, Terragrunt can
successfully download and run the modules without incorporating unwanted
example baggage. This affects the following files:

*   global/terraform.tfvars
*   vpc/terraform.tfvars

For similar reasons, we comment out the `backend.tf` files, also
*in their entirety*. SDG's Terraform infrastructure definitions
use a top-level `terragrunt.hcl` file that derives the key for the
backend's remote state based on the infrastructure definition directory's
relative path from the repo's top level, using the following pattern.

    `key    = "terraform/${path_relative_to_include()}/terraform.tfstate"`

This obviates the need to maintain `backend.tf` files in each infrastructure
directory. SDG has 200+ infrastructure directories per AWS account, so this
matters to us! The files affected are:

*   example-service/backend.tf (not used by SDG)
*   global/backend.tf
*   modules/bootstrap/backend.tf (not used by SDG)
*   vpc/backend.tf

Note that **all** `backend.tf` files have been commented out. For
consistency, this includes `backend.tf` files in two directories not
used by SDG.